Modified the rx_poll to loop based on reading the completion status pidx and current cidx
The previous version of the driver's rx_poll contained a while loop that checked for color matching in order to receive/process incoming packets.  The QDMA maintains completion entries that contain a color marking bit to help identify batches of transfers of data between the card and the host.  The color flips at the end of a batch.  This version of the driver instead relies on the completion status from the QDMA rather than color information.  Specifically, the driver reads the producer index (pidx), which indicates how many descriptors have been written.  This version of the driver reads/processes incoming packets until it's consumer index (cidx) matches the pidx.