Refactor interrupts
Currently the vector ID and queue ID are the same and the ID starts a 0. User
interrupt are allocated after queue interrupts. This results in the user
interrupt ending up at different locations depending on the number of vectors
available in the system. This is a problem as the user interrupt vector is
hardcoded in the shell and not configurable from SW.

This patch reorders the interrupt such that it matches the QDMA reference
driver:

0. mailbox interrupt (optional - defaults to unused)
1. user interrupt (optional - defaults to unused)
2. error interrupt on PF0
3. queue 0 interrupt
4. queue 1 interrupt
...

Hence, if mailbox interrupts is disabled, user interrupt will end up a interrupt
vector 0 and if enabled at interrupt vector 1.

Signed-off-by: Lars Munch <lars@segv.dk>