Update onic_hardware.c
This is a fix for incorrect resets being toggled when resetting CMAC, which could lead to the CMAC RX status being not ready (0xC0 rather than 0x3).   The values written to `SYSCFG_OFFSET_SHELL_RESET` map each bit to a separate reset signal.  To match the RTL encoding, these values need to be updated.      
The relevant details are "for CMAC0, the value should be 0x10, instead of 0x2; For CMAC1, the value should be 0x100, instead of 0x4. Because in the hardware (https://github.com/Xilinx/open-nic-shell/blob/main/src/open_nic_shell.sv#L385 ), cmac_rstn[0] and cmac_rstn[1] are actually assigned by shell_rstn[4] and shell_rstn[8], respectively."  
Thanks to Henry for catching this issue and finding a solution.