Fix to compile on linux kernel 6.4: Rename FIELD_GET to BITFIELD_GET
The macro FIELD_GET is already defined in linux/bitfield.h, that is
somehow included when compiling the onic driver on linux 6.4. This
causes the following error:

open-nic-driver/onic_common.h:64: error: "FIELD_GET" redefined [-Werror]
   64 | #define FIELD_GET(mask, reg)    ((reg & mask) >> FIELD_SHIFT(mask))

Unfortunately, the FIELD_GET macro that is already defined in
linux/bitfield.h is using the 'typeof' operator, and the 'typeof'
operator cannot be applied to bit-field members:
https://en.cppreference.com/w/cpp/language/bit_field
https://en.cppreference.com/w/c/language/typeof

The onic driver is making heavy use of bit-fields, thus, we cannot use
the already defined FIELD_GET macro without a heavy code refactoring.

This commit simply renames the FIELD_GET macro defined in the onic
driver to BITFIELD_GET to avoid the macro redefinition error.

Signed-off-by: Vangelis Tasoulas <vangelis@meta.com>